Create README.md
Speech to Text & ESP32 Tool
In this post I have added two tasks..
 
The First Task is website for covnverts Voice To Text and i used for this task HTML & JS. 

THE SECOND TASK is how we can to turn on the lighting. 
so the task writing How it's work ESP32.